Added quick auto test.

+++ src/org/usfirst/frc/team869/robot/commands/AutoCrossDefense.java
@@ -12,7 +12,7 @@ import org.usfirst.frc.team869.robot.commands.DriveDistanceCommand;
 public class AutoCrossDefense extends CommandGroup {
     
 	private double speed = .5;
-	private int distance = 50;//TODO find a good distance to travel
+	private double distance = 50;//TODO find a good distance to travel
 	
     public  AutoCrossDefense() {
